#25 - Revamped lecture on web applications with Spring.
Moved the old version of the script into a new folder so that it'll be still available. Added a link to the old version from the index site.

Reworked the lecture script to focus more on the way web applications work in general and how Spring MVC allows to work with the fundamental concepts of the web.